requirements.txt
aiohttp==3.6.2
python-dotenv==0.10.3
requests==2.22.0
